reorganized plotting classes into a new module
